Author

+++ SimpleKit/Controllers/WikiController.php
@@ -22,6 +22,10 @@ class WikiController extends BaseController {
         $this->wikitag = new Wikitags();
     }
 
+    public static function hasPermissionTo (int $id, Wiki $wiki) {
+        return $wiki->getById($id)[0]['authorID'] === $_SESSION['user_id'] || $_SESSION['user_role'] === 'admin';
+    }
+
     public function indexUserWikis() {
         if (!isset($_SESSION['user_id'])) {
             redirect('/login');
@@ -38,10 +42,6 @@ class WikiController extends BaseController {
     }
 
     public function index() {
-        if (!isset($_SESSION['user_id'])) {
-            redirect('/login');
-            exit();
-        }
         // Fetch all users using the wiki
         try {
             $wikis = $this->wiki->raw("SELECT w.id, w.title, w.content, c.name AS category, u.username AS author, GROUP_CONCAT(t.name) AS tags FROM wiki w JOIN user u ON w.authorID = u.id JOIN category c ON w.categoryID = c.id LEFT JOIN wiki_tags wt ON w.id = wt.wikiID LEFT JOIN tag t ON wt.tagID = t.id GROUP BY w.id, w.title, w.content, c.name, u.username ORDER BY w.created_at DESC;");
@@ -98,36 +98,43 @@ class WikiController extends BaseController {
     }
 
     public function edit(Request $request, $id) {
-       try {
-         // Fetch a specific wik by ID using the wiki
-         $this->wiki->updateById($id, ['title' => $request->getPostData("title"), 'content' => $request->getPostData("content"),'categoryID' => $request->getPostData("categoryID")]);
-         //delete old tags
-         $tagIdsToDelete = $this->wikitag->raw("SELECT tagID FROM wiki_tags WHERE wikiID = :wikiId", ['wikiId' => $id]);
-         foreach ($tagIdsToDelete as $tagID) $this->tag->deleteById($tagID['tagID']);
-         // add new tags
-         $lastInsertedTagIDs = [];
-             $tags = $request->getPostData("tags");
-             foreach ($tags as $value) $lastInsertedTagIDs[] = $this->tag->create(['name' => $value]);
-             //assign tags to wiki in wiki_tags pivot table
-             foreach ($tags as $index => $value) $this->wikitag->create(['wikiID' => $id, 'tagID' => $lastInsertedTagIDs[$index][0]]);
-             echo json_encode(["status" => "success", "message" => "Wiki edited successfully"]);
-       } catch (\Exception $e) {
-        echo json_encode(["status" => "insert", "message" => "There was an error editing the wiki"]);
-       }
+      if (WikiController::hasPermissionTo($id, $this->wiki)) {
+        try {
+            // Fetch a specific wik by ID using the wiki
+            $this->wiki->updateById($id, ['title' => $request->getPostData("title"), 'content' => $request->getPostData("content"),'categoryID' => $request->getPostData("categoryID")]);
+            //delete old tags
+            $tagIdsToDelete = $this->wikitag->raw("SELECT tagID FROM wiki_tags WHERE wikiID = :wikiId", ['wikiId' => $id]);
+            foreach ($tagIdsToDelete as $tagID) $this->tag->deleteById($tagID['tagID']);
+            // add new tags
+            $lastInsertedTagIDs = [];
+                $tags = $request->getPostData("tags");
+                foreach ($tags as $value) $lastInsertedTagIDs[] = $this->tag->create(['name' => $value]);
+                //assign tags to wiki in wiki_tags pivot table
+                foreach ($tags as $index => $value) $this->wikitag->create(['wikiID' => $id, 'tagID' => $lastInsertedTagIDs[$index][0]]);
+                echo json_encode(["status" => "success", "message" => "Wiki edited successfully"]);
+          } catch (\Exception $e) {
+           echo json_encode(["status" => "insert", "message" => "There was an error editing the wiki"]);
+          }
+      } else {
+        echo json_encode(["status" => "permission", "message" => "User doesnt have permission to perform this action"]);
+      }
     }
 
 
     public function destroy($id) {
-
-        try {
-            // Delete a specific wik by ID using the wiki
-        $tagIdsToDelete = $this->wikitag->raw("SELECT tagID FROM wiki_tags WHERE wikiID = :wikiId", ['wikiId' => $id]);
-        foreach ($tagIdsToDelete as $tagID) $this->tag->deleteById($tagID['tagID']);
-        $this->wiki->deleteById($id);
-        // Redirect back to the index page with a success message (or handle differently based on your needs)
-        echo json_encode(["status" => "success", "message" => "Wiki Deleted successfully"]);
-        } catch (\Exception $e) {
-        echo json_encode(["status" => "success", "message" => "There was an error deleting the wiki"]);
+        if (WikiController::hasPermissionTo($id, $this->wiki)) {
+            try {
+                // Delete a specific wik by ID using the wiki
+            $tagIdsToDelete = $this->wikitag->raw("SELECT tagID FROM wiki_tags WHERE wikiID = :wikiId", ['wikiId' => $id]);
+            foreach ($tagIdsToDelete as $tagID) $this->tag->deleteById($tagID['tagID']);
+            $this->wiki->deleteById($id);
+            // Redirect back to the index page with a success message (or handle differently based on your needs)
+            echo json_encode(["status" => "success", "message" => "Wiki Deleted successfully"]);
+            } catch (\Exception $e) {
+            echo json_encode(["status" => "success", "message" => "There was an error deleting the wiki"]);
+            }
+        } else {
+            echo json_encode(["status" => "permission", "message" => "User doesnt have permission to perform this action"]);
         }
     }
